refactor: split search implementation into separate modules
- Extract SearchAPI to search-api.ts for Pagefind wrapper
- Extract SearchUI to search-ui.ts for UI implementation
- Dynamic imports reduce initial bundle size further
- Search modules only load when modal opens (true lazy loading)
- Update tests to handle async modal operations
- Fix all lint issues and update snapshots
