+++ src/layouts/Layout.spec.ts
@@ -526,7 +526,7 @@ describe("Layout search modal (customSearch)", () => {
     expect(dialog?.close).toHaveBeenCalled();
   });
 
-  it("closes modal when clicking outside dialog frame", ({ expect }) => {
+  it("closes modal when clicking outside dialog frame", async ({ expect }) => {
     const openBtn =
       document.querySelector<HTMLButtonElement>("[data-open-modal]");
     const dialog = document.querySelector<HTMLDialogElement>("dialog");
@@ -534,6 +534,9 @@ describe("Layout search modal (customSearch)", () => {
     // Open the modal
     openBtn?.click();
     expect(dialog?.open).toBe(true);
+    
+    // Wait for async operations to complete
+    await new Promise((resolve) => setTimeout(resolve, 10));
 
     // Click outside the dialog frame (on body)
     const clickEvent = new window.MouseEvent("click", {
